creaed a common function to get queryVariables for create and review queries
