+++ console/console-init/ui/src/modules/address-space/dialogs/CreateMessagingProject/MessagingProjectReview.tsx
@@ -11,7 +11,8 @@ import { AddressSpaceReview } from "modules/address-space/components";
 import { IMessagingProject, IExposeEndPoint } from "./CreateMessagingProject";
 import {
   EndPointProtocol,
-  TlsCertificateType
+  TlsCertificateType,
+  getQueryVariableForCreateAddressSpace
 } from "modules/address-space/utils";
 
 interface IMessagingProjectReviewProps {
@@ -34,75 +35,8 @@ export const MessagingProjectReview: React.FunctionComponent<IMessagingProjectRe
   } = projectDetail || {};
 
   const queryVariable = {
-    variables: {
-      as: {
-        metadata: {
-          name: name,
-          namespace: namespace
-        },
-        spec: {
-          plan: plan ? plan.toLowerCase() : "",
-          type: type ? type.toLowerCase() : "",
-          authenticationService: {
-            name: authService
-          }
-        }
-      }
-    }
+    variables: getQueryVariableForCreateAddressSpace(projectDetail)
   };
-
-  if (customizeEndpoint) {
-    const { certValue, privateKey, routesConf } = projectDetail;
-    const endpoints: IExposeEndPoint[] = [];
-    if (protocols && protocols.length > 0) {
-      protocols.map((protocol: string) => {
-        const endpoint: IExposeEndPoint = { service: "messaging" };
-        if (protocol === EndPointProtocol.AMQPS) {
-          endpoint.name = "messaging";
-        } else if (protocol === EndPointProtocol.AMQP_WSS) {
-          endpoint.name = "messaging-wss";
-        }
-        if (tlsCertificate) {
-          endpoint.certificate = {
-            provider: tlsCertificate
-          };
-          if (
-            tlsCertificate === TlsCertificateType.UPLOAD_CERT &&
-            certValue &&
-            certValue.trim() !== "" &&
-            privateKey &&
-            privateKey.trim() !== ""
-          ) {
-            endpoint.certificate = {
-              ...endpoint.certificate,
-              tlsKey: privateKey?.trim(),
-              tlsCert: certValue?.trim()
-            };
-          }
-        }
-        if (addRoutes) {
-          endpoint.expose = { type: "route", routeServicePort: protocol };
-          const routeConf = routesConf?.filter(
-            conf => conf.protocol === protocol
-          );
-          if (routeConf && routeConf.length > 0) {
-            if (routeConf[0].hostname && routeConf[0].hostname.trim() !== "") {
-              endpoint.expose.routeHost = routeConf[0].hostname.trim();
-            }
-            if (
-              routeConf[0].tlsTermination &&
-              routeConf[0].tlsTermination.trim() !== ""
-            ) {
-              endpoint.expose.routeTlsTermination = routeConf[0].tlsTermination;
-            }
-          }
-        }
-        endpoints.push(endpoint);
-        return endpoint;
-      });
-    }
-    Object.assign(queryVariable.variables.as.spec, { endpoints: endpoints });
-  }
 
   const { data, loading } = useQuery(
     ADDRESS_SPACE_COMMAND_REVIEW_DETAIL,
